[clang][cas] Move caching ActionControllers into separate files
Move the IncludeTreeActionController and CASFSActionController into
their own files.
